backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.11.5
pydantic-settings==2.1.0

# LlamaIndex Core & Components
llama-index-core
llama-index-llms-groq
llama-index-embeddings-ollama
llama-index-vector-stores-postgres
llama-index-readers-file

# Reranking (needed for SentenceTransformerRerank)
torch
sentence-transformers
transformers

# Vector Store & Search
psycopg2-binary
pgvector
sqlalchemy
rank-bm25

# File Processing (additional dependencies for file readers)
pypdf==4.3.1
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.2

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1
typing-extensions==4.12.2
nest-asyncio==1.6.0

# Logging & Monitoring
structlog==23.2.0

# HTTP Client
httpx==0.25.2

# Password Hashing
passlib[argon2]
argon2-cffi
